+++ Distributions/Discrete/DiscreteUniform.cs
@@ -32,7 +32,7 @@ namespace Chinchillada.Distributions
             return _support.Contains(variable).ToBinary();
         }
 
-        double IWeightedDistribution<T>.Weight(T item)
+        float IWeightedDistribution<T>.Weight(T item)
         {
             return Weight(item);
         }
