+++ Distributions/Discrete/Projected.cs
@@ -46,7 +46,7 @@ namespace Chinchillada.Distributions
         public IEnumerable<R> Support() => _weights.Keys;
 
         public int Weight(R variable) => _weights.GetValueOrDefault(variable);
-        double IWeightedDistribution<R>.Weight(R item)
+        float IWeightedDistribution<R>.Weight(R item)
         {
             return Weight(item);
         }
