+++ Distributions/Discrete/WeightedInteger.cs
@@ -109,7 +109,7 @@
             return _weights.ElementAtOrDefault(variable);
         }
 
-        double IWeightedDistribution<int>.Weight(int item)
+        float IWeightedDistribution<int>.Weight(int item)
         {
             return Weight(item);
         }
